# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Minecraft Hosting Panel API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginRequest {
  login: string;
  password: string;
}

export type AuthUserRole = (typeof AuthUserRole)[keyof typeof AuthUserRole];

export const AuthUserRole = {
  admin: "admin",
  user: "user",
} as const;

export interface AuthUser {
  id: string;
  username: string;
  email: string;
  role: AuthUserRole;
  createdAt: string;
}

export interface UpdateProfileRequest {
  username?: string;
  email?: string;
  password?: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export type ServerStatus = (typeof ServerStatus)[keyof typeof ServerStatus];

export const ServerStatus = {
  starting: "starting",
  running: "running",
  stopping: "stopping",
  stopped: "stopped",
  crashed: "crashed",
} as const;

export type ServerServerType =
  (typeof ServerServerType)[keyof typeof ServerServerType];

export const ServerServerType = {
  vanilla: "vanilla",
  paper: "paper",
  spigot: "spigot",
  forge: "forge",
  fabric: "fabric",
  bungeecord: "bungeecord",
} as const;

export interface Server {
  id: string;
  name: string;
  description?: string;
  status: ServerStatus;
  version: string;
  port: number;
  memory: number;
  disk: number;
  nodeId: string;
  userId: string;
  createdAt: string;
  updatedAt: string;
  playerCount?: number;
  maxPlayers?: number;
  javaVersion?: string;
  serverType: ServerServerType;
}

export type CreateServerRequestServerType =
  (typeof CreateServerRequestServerType)[keyof typeof CreateServerRequestServerType];

export const CreateServerRequestServerType = {
  vanilla: "vanilla",
  paper: "paper",
  spigot: "spigot",
  forge: "forge",
  fabric: "fabric",
  bungeecord: "bungeecord",
} as const;

export interface CreateServerRequest {
  name: string;
  description?: string;
  version: string;
  port: number;
  memory: number;
  disk: number;
  nodeId: string;
  userId: string;
  maxPlayers?: number;
  javaVersion?: string;
  serverType: CreateServerRequestServerType;
}

export type PowerActionRequestAction =
  (typeof PowerActionRequestAction)[keyof typeof PowerActionRequestAction];

export const PowerActionRequestAction = {
  start: "start",
  stop: "stop",
  restart: "restart",
  kill: "kill",
} as const;

export interface PowerActionRequest {
  action: PowerActionRequestAction;
}

export interface CommandRequest {
  command: string;
}

export interface ServerStats {
  serverId: string;
  cpuPercent: number;
  memoryUsed: number;
  memoryLimit: number;
  diskUsed: number;
  diskLimit: number;
  networkRx: number;
  networkTx: number;
  uptime: number;
  playerCount: number;
}

export interface LogsResponse {
  serverId: string;
  logs: string[];
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  user: "user",
} as const;

export interface User {
  id: string;
  username: string;
  email: string;
  role: UserRole;
  createdAt: string;
}

export type CreateUserRequestRole =
  (typeof CreateUserRequestRole)[keyof typeof CreateUserRequestRole];

export const CreateUserRequestRole = {
  admin: "admin",
  user: "user",
} as const;

export interface CreateUserRequest {
  username: string;
  email: string;
  password: string;
  role?: CreateUserRequestRole;
}

export type NodeStatus = (typeof NodeStatus)[keyof typeof NodeStatus];

export const NodeStatus = {
  online: "online",
  offline: "offline",
  maintenance: "maintenance",
} as const;

export interface Node {
  id: string;
  name: string;
  fqdn: string;
  port: number;
  memory: number;
  disk: number;
  status: NodeStatus;
  location?: string;
  createdAt: string;
  serversCount?: number;
}

export interface CreateNodeRequest {
  name: string;
  fqdn: string;
  port: number;
  memory: number;
  disk: number;
  location?: string;
}

export interface UpdateServerRequest {
  name?: string;
  description?: string;
  memory?: number;
  disk?: number;
  maxPlayers?: number;
  port?: number;
}

export interface FileEntry {
  name: string;
  path: string;
  isDir: boolean;
  size: number;
  updatedAt: string;
}

export interface FileListResponse {
  path: string;
  entries: FileEntry[];
}

export interface FileReadResponse {
  path: string;
  content: string;
  size: number;
  updatedAt: string;
}

export interface FileWriteRequest {
  content: string;
}

export interface ServerPlugin {
  id: string;
  serverId: string;
  name: string;
  version: string;
  description?: string;
  author?: string;
  filename: string;
  fileSize: number;
  enabled: boolean;
  createdAt: string;
}

export interface InstallPluginRequest {
  name: string;
  version?: string;
}

export interface Subdomain {
  id: string;
  serverId: string;
  subdomain: string;
  targetPort: string;
  createdAt: string;
}

export interface CreateSubdomainRequest {
  subdomain: string;
}

export type GetServerLogsParams = {
  lines?: number;
};

export type ListServerFilesParams = {
  path?: string;
};

export type DeleteServerFileParams = {
  path: string;
};

export type ReadServerFileParams = {
  path: string;
};

export type WriteServerFileParams = {
  path: string;
};

export type CreateServerDirectoryBody = {
  path: string;
};

export type ToggleServerPluginBody = {
  enabled: boolean;
};