# lib/api-zod/src/generated/types/serverPlugin.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Minecraft Hosting Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface ServerPlugin {
  id: string;
  serverId: string;
  name: string;
  version: string;
  description?: string;
  author?: string;
  filename: string;
  fileSize: number;
  enabled: boolean;
  createdAt: Date;
}